requirements.txt
torch>=1.12
torchvision>=0.13
opencv-python
numpy
matplotlib
pycocotools
scikit-learn
scipy
tqdm
pillow
lxml 
pyYAML
pycocotools
